Upload Image with Cloudinary

+++ pages/api/products/[id].js
@@ -6,9 +6,10 @@ import Product from '../../../models/Product'
 export default async function handler(req, res) {
   const { 
       method,  
-      query: {id},
+      query: {id}, cookies
     } = req;
 
+  const token = cookies.token;
   dbConnect()
 
   if (method === 'GET') {
@@ -22,6 +23,10 @@ export default async function handler(req, res) {
   }
 
   if (method === 'PUT') {
+    if (!token || token !== process.env.token)
+    {
+        return res.status(401).json("Not authenticated!")
+    }
     try {
 
       const product = await Product.findByIdAndUpdate(id, req.body, {
@@ -34,6 +39,10 @@ export default async function handler(req, res) {
 }
 
 if (method === 'DELETE') {
+    if (!token || token !== process.env.token)
+    {
+        return res.status(401).json("Not authenticated!")
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!")
